Refactor movie data handling to use Long for budget and revenue, and introduce MovieDataAssembler for improved data assembly

+++ src/main/java/com/mgrunt/movies/domain/dtos/movie/MovieDetailsResponse.java
@@ -22,7 +22,7 @@ public record MovieDetailsResponse(
         List<TmdbGenreResponse> genres,
         String originalLanguage,
         Boolean adult,
-        Integer budget,
+        Long budget,
         Integer revenue,
         String tagline,
         String trailerUrl,
